Extract ContextAwareException from LocationAwareException
