+++ subprojects/core/src/main/java/org/gradle/initialization/exception/InitializationException.java
@@ -22,7 +22,9 @@ import org.gradle.internal.exceptions.Contextual;
 @Contextual
 public class InitializationException extends GradleException {
 
+    public static final String MESSAGE = "Gradle could not start your build.";
+
     public InitializationException(Throwable cause) {
-        super("Gradle could not start your build.", cause);
+        super(MESSAGE, cause);
     }
 }
