+++ subprojects/core/src/main/java/org/gradle/internal/exceptions/LocationAwareException.java
@@ -16,22 +16,16 @@
 package org.gradle.internal.exceptions;
 
 import org.apache.commons.lang.StringUtils;
-import org.gradle.api.GradleException;
 import org.gradle.groovy.scripts.ScriptSource;
 import org.gradle.initialization.BuildClientMetaData;
 import org.gradle.internal.logging.text.StyledTextOutput;
 import org.gradle.internal.scan.UsedByScanPlugin;
-import org.gradle.util.TreeVisitor;
-
-import javax.annotation.Nullable;
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * A {@code LocationAwareException} is an exception which can be annotated with a location in a script.
  */
 @UsedByScanPlugin
-public class LocationAwareException extends GradleException implements FailureResolutionAware {
+public class LocationAwareException extends ContextAwareException implements FailureResolutionAware {
     private final String sourceDisplayName;
     private final Integer lineNumber;
 
@@ -40,9 +34,9 @@ public class LocationAwareException extends GradleException implements FailureRe
     }
 
     public LocationAwareException(Throwable cause, String sourceDisplayName, Integer lineNumber) {
+        super(cause);
         this.sourceDisplayName = sourceDisplayName;
         this.lineNumber = lineNumber;
-        initCause(cause);
     }
 
     /**
@@ -59,6 +53,7 @@ public class LocationAwareException extends GradleException implements FailureRe
      *
      * @return The location description. May return null.
      */
+    @Override
     public String getLocation() {
         if (sourceDisplayName == null) {
             return null;
@@ -108,70 +103,4 @@ public class LocationAwareException extends GradleException implements FailureRe
         }
     }
 
-    /**
-     * Returns the reportable causes for this failure.
-     *
-     * @return The causes. Never returns null, returns an empty list if this exception has no reportable causes.
-     */
-    public List<Throwable> getReportableCauses() {
-        final List<Throwable> causes = new ArrayList<Throwable>();
-        visitCauses(getCause(), new TreeVisitor<Throwable>() {
-            @Override
-            public void node(Throwable node) {
-                causes.add(node);
-            }
-        });
-        return causes;
-    }
-
-    /**
-     * Visits the reportable causes for this failure.
-     */
-    public void visitReportableCauses(TreeVisitor<? super Throwable> visitor) {
-        visitor.node(this);
-        if (this.getCause() != null) {
-            visitCauses(this.getCause(), visitor);
-        }
-    }
-
-    private void visitCauses(Throwable t, TreeVisitor<? super Throwable> visitor) {
-        if (t instanceof MultiCauseException) {
-            MultiCauseException multiCauseException = (MultiCauseException) t;
-            List<? extends Throwable> causes = multiCauseException.getCauses();
-            if (!causes.isEmpty()) {
-                visitor.startChildren();
-                for (Throwable cause : causes) {
-                    visitContextual(cause, visitor);
-                }
-                visitor.endChildren();
-            }
-        } else if (t.getCause() != null) {
-            visitor.startChildren();
-            visitContextual(t.getCause(), visitor);
-            visitor.endChildren();
-        }
-    }
-
-    private void visitContextual(Throwable t, TreeVisitor<? super Throwable> visitor) {
-        Throwable next = findNearestContextual(t);
-        if (next != null) {
-            // Show any contextual cause recursively
-            visitor.node(next);
-            visitCauses(next, visitor);
-        } else {
-            // Show the direct cause of the last contextual cause only
-            visitor.node(t);
-        }
-    }
-
-    @Nullable
-    private Throwable findNearestContextual(@Nullable Throwable t) {
-        if (t == null) {
-            return null;
-        }
-        if (t.getClass().getAnnotation(Contextual.class) != null || t instanceof MultiCauseException) {
-            return t;
-        }
-        return findNearestContextual(t.getCause());
-    }
 }
