+++ subprojects/launcher/src/main/java/org/gradle/launcher/cli/ExceptionReportingAction.java
@@ -18,8 +18,9 @@ package org.gradle.launcher.cli;
 import org.gradle.api.Action;
 import org.gradle.initialization.ReportedException;
 import org.gradle.initialization.exception.InitializationException;
-import org.gradle.internal.exceptions.LocationAwareException;
+import org.gradle.internal.exceptions.ContextAwareException;
 import org.gradle.internal.logging.LoggingOutputInternal;
+import org.gradle.internal.service.ServiceCreationException;
 import org.gradle.launcher.bootstrap.ExecutionListener;
 
 public class ExceptionReportingAction implements Action<ExecutionListener> {
@@ -44,8 +45,11 @@ public class ExceptionReportingAction implements Action<ExecutionListener> {
         } catch (ReportedException e) {
             // Exception has already been reported
             executionListener.onFailure(e);
+        } catch (ServiceCreationException e) {
+            reporter.execute(new ContextAwareException(new InitializationException(e)));
+            executionListener.onFailure(e);
         } catch (Throwable t) {
-            reporter.execute(new LocationAwareException(new InitializationException(t), (String) null, null));
+            reporter.execute(t);
             executionListener.onFailure(t);
         }
     }
